function/requirements.txt
azure-functions
azure-eventgrid>=4.16.0
azure-identity>=1.15.0
